refactor: structure the lib according to guidelines
