+++ lib/src/rabbit.dart
@@ -1,0 +1,30 @@
+part of rabbit_converter;
+
+abstract class Rabbit {
+  /// Converts Myanmar Unicode Encoded Strings to Zawgyi Encoding.
+  static String uni2zg(String input) {
+    return _replaceWithRule(Rules.uni2zg, input);
+  }
+
+  /// Converts Zawgyi Encoded Strings to Myanmar Unicode Encoding.
+  static String zg2uni(String input) {
+    return _replaceWithRule(Rules.zg2uni, input);
+  }
+
+  static String _replaceWithRule(
+    List<Map<String, String>> rules,
+    String input,
+  ) {
+    var output = input;
+
+    for (final rule in rules) {
+      output = output.replaceAllMapped(RegExp(rule['from']), (match1) {
+        return rule['to'].replaceAllMapped(RegExp(r'#(\d{1})'), (match2) {
+          return match1.group(int.parse(match2.group(1))) ?? '';
+        });
+      });
+    }
+
+    return output;
+  }
+}
